% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gistoolS3G5.R
\name{gistoolS3G5}
\alias{gistoolS3G5}
\title{Gradient analysis of WB}
\usage{
gistoolS3G5()
}
\arguments{
\item{WBData}{Tables of gradient polygons}
}
\value{
Vertical profiles and Monod growth kinetics in various plots
}
\description{
Uses Model data for WB divided into five gradient polygons. Model scenarios are "Baseline", "N-30%" and "P-30%"
}
