% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HF.R
\name{HF_GIS_tool}
\alias{HF_GIS_tool}
\title{Gradient analysis of HF}
\usage{
HF_GIS_tool()
}
\arguments{
\item{HFData}{Tables of gradient polygons}
}
\value{
Vertical profiles and Monod growth kinetics in various plots
}
\description{
Uses Model data for HF divided into five gradient polygons. Model scenarios are "Baseline", "N-30%" and "P-30%"
}
