% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VF.R
\name{VF_GIS_tool}
\alias{VF_GIS_tool}
\title{Gradient analysis of VF}
\usage{
VF_GIS_tool()
}
\arguments{
\item{VFData}{Tables of gradient polygons}
}
\value{
Vertical profiles and Monod growth kinetics in various plots
}
\description{
Uses Model data for VF divided into five gradient polygons. Model scenarios are "Baseline", "N-30%" and "P-30%"
}
